+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `cool` operation mode to `AtwDevice` zones.
+- Add zone flow and flow return temperature read.
 
 ## [2.4.1] - 2020-03-29
 ### Fixed